telc-b1-speaking-trainer/requirements.txt
openai = 1.60.2
python-dotenv = 1.0.1
unittest2 = 1.1.0